Fixed bugs when saving asset
Several bugs when saving asset were found and fixed in this commit,
including the bug of incorrect file format and incorrect method usage in request mapper.

+++ web-controller/src/main/java/com/oasis/web_controller/mapper/request/RequestsRequestMapper.java
@@ -22,7 +22,7 @@ public class RequestsRequestMapper {
             requestModel.set_id(requestObject.get_id());
             requestModel.setUsername(username);
             requestModel.setSku(requestObject.getSku());
-            requestModel.setQuantity((int) requestObject.getQuantity());
+            requestModel.setQuantity(requestObject.getQuantity());
             requestModel.setStatus(requestObject.getStatus());
             requestModel.setRequestNote(requestObject.getRequestNote());
             requestModel.setTransactionNote(requestObject.getTransactionNote());
